requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
branca==0.8.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.0
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.1.1
django-active-link==0.2.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-sesame==3.2.2
django-taggit==6.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
fastapi==0.115.0
Flask==3.1.0
folium==0.19.3
fonttools==4.55.0
frozendict==2.4.6
frozenlist==1.5.0
googlemaps==4.10.0
h11==0.16.0
html5lib==1.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.4.2
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
MouseInfo==0.1.3
multidict==6.1.0
multitasking==0.0.11
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
opencage==3.0.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
peewee==3.17.7
pefile==2023.2.7
phonenumbers==8.13.53
pillow==10.4.0
pip-chill==1.0.3
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.9.0
PyMsgBox==1.0.9
pyOpenSSL==25.0.0
pyparsing==3.2.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.1.1
pytest-metadata==3.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytube==15.0.0
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.0
redis==5.1.0
referencing==0.36.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
reverse_geocoder==1.5.1
rpds-py==0.23.1
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.32.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.6
speedtest-cli==2.1.3
sqlparse==0.5.1
starlette==0.38.6
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
xyzservices==2024.9.0
yarl==1.18.3
yfinance==0.2.49
yt-dlp==2024.11.18
